requirements.txt
opencv-python
numpy==1.23.5
trimesh
plyfile
pyyaml
scikit-image
scikit-learn
chumpy
tensorboard
kornia
loguru
pycocotools
yacs
lmdb
fire
setuptools==59.5.0
einops
tqdm
ipython
gym
transformers==4.34.1
OmegaConf
matplotlib
gradio
diffusers==0.19.3
rembg
segment_anything
jaxtyping
imageio
iopath
timm
open3d
